+++ NEWS.md
@@ -1,3 +1,5 @@
+# surveytable (development version)
+
 # surveytable 0.9.5
 
 * Another public use data file for use in examples: `rccsu2018`.